requirements.txt
black
numpy
openai
pdfplumber
pymupdf
python-docx
ruff
sendgrid
streamlit
tiktoken
docx
pandas
lxml_html_clean
lxml==4.9.3
newspaper3k
openpyxl
playwright==1.43.0
rapidfuzz 
msal
asyncio
